requirements.txt
annotated-types==0.7.0
asgiref==3.9.1
cffi==1.17.1
contextlib2==21.6.0
cryptography==45.0.6
Django==5.2.5
django-cors-headers==4.7.0
django-ninja==1.4.3
django-ninja-extra==0.30.1
django-ninja-jwt==5.3.7
django-user-agents==0.4.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
injector==0.22.0
mssql-django==1.6
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.10.1
pyodbc==5.2.0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
sqlparse==0.5.3
typing-inspection==0.4.1
typing_extensions==4.14.1
ua-parser==1.0.1
ua-parser-builtins==0.18.0.post1
user-agents==2.2.0
